refactor: 사이드바·도움말 말풍선 렌더 분리 #88
role-sidebar 217 → 118 (항목 렌더·아이콘 맵을 sidebar-item으로)
support-thread 201 → 96 (말풍선을 support-turn으로)
