+++ src/features/shell/components/role-sidebar.tsx
@@ -1,58 +1,18 @@
 "use client";
 
-import {
-  CalendarDays,
-  CalendarRange,
-  Columns3,
-  CreditCard,
-  Folder,
-  HardDrive,
-  LayoutDashboard,
-  type LucideIcon,
-  Megaphone,
-  Search,
-  Settings,
-  UserRound,
-  UserRoundCheck,
-  Users,
-  Video,
-} from "lucide-react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { toast } from "sonner";
 
 import { ZLogo } from "@/components/icons/z-logo";
 import { ROLE_LABEL } from "@/constants/domain";
-import { cn } from "@/lib/utils";
 
 import { roleHome } from "../home";
-import type { NavIconName, NavItem, NavSection } from "../nav";
+import type { NavSection } from "../nav";
 import type { Viewer } from "../viewer";
-
-/**
- * 이름 → 아이콘. 구성 파일은 서버에서 읽히므로 실제 컴포넌트는 여기서 붙인다.
- * OWNER_NAV가 쓰는 이름만 채운다 — `Partial`이라 나머지(SYSTEM 전용 등)는 없어도 된다.
- */
-const NAV_ICON: Partial<Record<NavIconName, LucideIcon>> = {
-  dashboard: LayoutDashboard,
-  project: Folder,
-  search: Search,
-  calendar: CalendarDays,
-  notice: Megaphone,
-  meeting: Video,
-  room: CalendarRange,
-  board: Columns3,
-  people: Users,
-  me: UserRound,
-  members: UserRoundCheck,
-  billing: CreditCard,
-  storage: HardDrive,
-  setting: Settings,
-};
+import { SidebarItem } from "./sidebar-item";
 
 interface RoleSidebarProps {
   sections: NavSection[];
-  /** 하단 계정 줄 */
   /**
    * 지금 보고 있는 사람 — 이름·역할 배지·로고 링크가 이 값에서 나온다.
    *
@@ -137,81 +97,22 @@ export function RoleSidebar({ sections, user }: RoleSidebarProps) {
         ))}
       </nav>
 
-      <div className="border-border flex h-[49px] shrink-0 items-center gap-[7px] border-t px-[17.5px]">
-        <span className="bg-role-owner flex size-[21px] shrink-0 items-center justify-center rounded-full text-[10px] leading-none text-white">
-          {user.name.slice(0, 1)}
-        </span>
-        <span className="min-w-0 flex-1 truncate text-xs leading-[18px]">{user.name}</span>
-        <span className="bg-role-owner-surface text-role-owner shrink-0 rounded px-[5.25px] py-[1.75px] text-[9px] leading-[14px]">
-          {ROLE_LABEL[user.role]}
-        </span>
-      </div>
+      <AccountRow user={user} />
     </aside>
   );
 }
 
-function SidebarItem({ item, isCurrent }: { item: NavItem; isCurrent: boolean }) {
-  const Icon = NAV_ICON[item.icon] ?? LayoutDashboard;
-  const inner = (
-    <>
-      <Icon className="size-[14px] shrink-0" aria-hidden />
-      {/* 한글 글자가 상자 안에서 위쪽에 앉아 아이콘보다 떠 보인다 — 1px 내려 맞춘다 */}
-      <span className="min-w-0 flex-1 translate-y-px truncate text-[13px] leading-5">
-        {item.label}
-      </span>
-      {item.dot && (
-        <>
-          <span className="bg-destructive size-[6px] shrink-0 rounded-full" aria-hidden />
-          <span className="sr-only">안 읽음</span>
-        </>
-      )}
-      {item.badge !== undefined && (
-        <span className="bg-foreground text-background flex h-[17px] min-w-[17px] shrink-0 items-center justify-center rounded-full px-[3.5px] text-[10px] leading-none font-semibold tabular-nums">
-          {item.badge}
-        </span>
-      )}
-    </>
-  );
-
-  const shape = "flex h-[34px] items-center gap-[8.75px] rounded-md px-[10.5px] transition-colors";
-
-  /*
-    ⚠️ 아직 없는 화면은 링크로 두지 않는다 — 누르면 404가 뜬다.
-       대신 **색은 다른 메뉴와 똑같이** 둔다. 흐리게 처리하면 완성된 화면을 볼 때
-       사이드바가 절반쯤 죽은 것처럼 보인다.
-       누르면 조용히 아무 일도 안 일어나는 대신 준비 중이라고 알린다(§정직성).
-  */
-  if (!item.isReady) {
-    return (
-      <button
-        type="button"
-        aria-disabled
-        // ⚠️ 토스트는 한 줄(220px)이라 짧게 쓴다 — 길면 잘린다(`sonner.tsx`)
-        onClick={() => toast(`${item.label}은 준비 중입니다`)}
-        // 색은 준비된 메뉴와 똑같이 — 평소 회색, 호버하면 글자가 진해진다
-        className={cn(
-          shape,
-          "text-muted-foreground hover:bg-foreground/5 hover:text-foreground focus-visible:ring-ring w-full text-left focus-visible:ring-2 focus-visible:outline-hidden",
-        )}
-      >
-        {inner}
-      </button>
-    );
-  }
-
+/** 하단 계정 줄 — 이름 첫 글자 · 이름 · 역할 배지 */
+function AccountRow({ user }: { user: Viewer }) {
   return (
-    <Link
-      href={item.href}
-      aria-current={isCurrent ? "page" : undefined}
-      className={cn(
-        shape,
-        "focus-visible:ring-ring focus-visible:ring-2 focus-visible:outline-hidden",
-        isCurrent
-          ? "bg-foreground text-background"
-          : "text-muted-foreground hover:bg-foreground/5 hover:text-foreground",
-      )}
-    >
-      {inner}
-    </Link>
+    <div className="border-border flex h-[49px] shrink-0 items-center gap-[7px] border-t px-[17.5px]">
+      <span className="bg-role-owner flex size-[21px] shrink-0 items-center justify-center rounded-full text-[10px] leading-none text-white">
+        {user.name.slice(0, 1)}
+      </span>
+      <span className="min-w-0 flex-1 truncate text-xs leading-[18px]">{user.name}</span>
+      <span className="bg-role-owner-surface text-role-owner shrink-0 rounded px-[5.25px] py-[1.75px] text-[9px] leading-[14px]">
+        {ROLE_LABEL[user.role]}
+      </span>
+    </div>
   );
 }
